fix test

+++ testing/dart/painting_test.dart
@@ -6,6 +6,7 @@ import 'dart:typed_data';
 import 'dart:ui';
 
 import 'package:litetest/litetest.dart';
+import 'package:vector_math/vector_math_64.dart';
 
 typedef CanvasCallback = void Function(Canvas canvas);
 
@@ -105,7 +106,7 @@ void main() {
   });
 
   test('ImageFilter.matrix defaults to FilterQuality.medium', () {
-    final Float32List data = Matrix4.identity().storage;
+    final Float64List data = Matrix4.identity().storage;
     expect(
       ImageFilter.matrix(data).toString(),
       'ImageFilter.matrix($data, FilterQuality.medium)',
